Extracting CF based paging logic to it's own type
