+++ sdk/core/azure-core-rest-cfextensions/src/main/java/com/azure/android/core/rest/cfextensions/PagedAsyncStream.java
@@ -5,8 +5,8 @@ package com.azure.android.core.rest.cfextensions;
 
 import com.azure.android.core.logging.ClientLogger;
 import com.azure.android.core.rest.PagedResponse;
+import com.azure.android.core.util.AsyncStream;
 
-import java9.util.concurrent.CompletableFuture;
 import java9.util.function.Function;
 
 /**
@@ -18,14 +18,15 @@ import java9.util.function.Function;
  */
 public final class PagedAsyncStream<T> extends ContinuablePagedAsyncStream<String, T, PagedResponse<T>> {
     /**
-     * Creates an instance of {@link PagedAsyncStream}. The constructor takes a {@code pageRetriever}.
-     * The {@code pageRetriever} returns a page of {@code T} when invoked with the id of the page to retrieve.
+     * Creates an instance of {@link PagedAsyncStream}. The constructor takes a {@code streamRetriever}.
+     * The {@code pageRetriever} returns {@code AsyncStream} of pages when invoked with the id of the page
+     * to retrieve.
      *
-     * @param pageRetriever Function that retrieves the page.
+     * @param streamRetriever The function to retrieve page stream backing this stream.
      * @param logger The logger to log.
      */
-    public PagedAsyncStream(Function<String, CompletableFuture<PagedResponse<T>>> pageRetriever,
+    public PagedAsyncStream(Function<String, AsyncStream<PagedResponse<T>>> streamRetriever,
                             ClientLogger logger) {
-        super(pageRetriever, pageId -> pageId != null, logger);
+        super(streamRetriever, logger);
     }
 }
